% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_utils.R
\name{breakpoint_distance}
\alias{breakpoint_distance}
\title{Breakpoint Distance Between Two States}
\usage{
breakpoint_distance(start_state, target_state)
}
\arguments{
\item{start_state}{Integer vector, first state}

\item{target_state}{Integer vector, second state}
}
\value{
Integer, the number of breakpoints
}
\description{
Counts the number of positions where consecutive elements differ by
more than 1 (breakpoints). Particularly effective for TopSpin puzzles
where operations shift blocks and flip prefixes.
}
\examples{
breakpoint_distance(1:5, 5:1)
breakpoint_distance(1:5, 1:5)
}
