% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu.R
\name{cayley_gpu_available}
\alias{cayley_gpu_available}
\title{Check if GPU acceleration is available}
\usage{
cayley_gpu_available()
}
\value{
Logical
}
\description{
Checks whether ggmlR is installed and Vulkan GPU is present.
}
\examples{
cayley_gpu_available()
}
