% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_processing.R
\name{find_best_match_state}
\alias{find_best_match_state}
\title{Find Best Match State}
\usage{
find_best_match_state(target_state, reachable_states, method = "manhattan",
  use_gpu = FALSE)
}
\arguments{
\item{target_state}{Integer vector, the target state}

\item{reachable_states}{Data frame or Arrow Table with V-columns}
}
\value{
Single-row data frame of the best matching state
}
\description{
Finds the state in a table that has the minimum Manhattan distance to
a target state. If multiple states tie, selects the one with the
smallest step number.
}
\keyword{internal}
