% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_utils.R
\name{generate_unique_states_df}
\alias{generate_unique_states_df}
\title{Generate Data Frame of Unique Random States}
\usage{
generate_unique_states_df(n, n_rows)
}
\arguments{
\item{n}{Integer, size of each permutation state}

\item{n_rows}{Integer, number of unique states to generate}
}
\value{
Data frame with n_rows rows and columns V1, V2, ..., Vn
}
\description{
Generates a data frame with unique random permutation states.
}
\examples{
set.seed(42)
df <- generate_unique_states_df(5, 10)
head(df)
}
