% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_utils.R
\name{manhattan_distance}
\alias{manhattan_distance}
\title{Manhattan Distance Between Two States}
\usage{
manhattan_distance(start_state, target_state)
}
\arguments{
\item{start_state}{Integer vector, first state}

\item{target_state}{Integer vector, second state}
}
\value{
Numeric, the Manhattan distance
}
\description{
Computes the sum of absolute differences between corresponding elements
of two permutation states.
}
\examples{
manhattan_distance(1:5, 5:1)
manhattan_distance(1:5, 1:5)
}
