% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_processing.R
\name{select_new_state}
\alias{select_new_state}
\title{Select New Bridge State}
\usage{
select_new_state(target_all, opposite_state)
}
\arguments{
\item{target_all}{Data frame of candidate states}

\item{opposite_state}{Integer vector, the state to be close to}
}
\value{
Integer vector of the selected state
}
\description{
Selects a new state from candidate states that is close to an opposite
state (by Manhattan distance). Randomly picks from the top 10 closest.
}
\keyword{internal}
