% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_utils.R
\name{generate_state}
\alias{generate_state}
\title{Generate Random Permutation State}
\usage{
generate_state(n)
}
\arguments{
\item{n}{Integer, the size of the permutation}
}
\value{
Integer vector representing a random permutation of 1:n
}
\description{
Generates a random permutation of integers from 1 to n.
}
\examples{
set.seed(42)
generate_state(10)
}
